requirements.txt
fastapi
uvicorn
python-multipart
openai-whisper
googletrans==4.0.0-rc1
ffmpeg-python